+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityEditModel.java
@@ -48,7 +48,5 @@ public class SwingEntityEditModel extends DefaultEntityEditModel<SwingEntityMode
 	public SwingEntityEditModel(EntityType entityType, EntityConnectionProvider connectionProvider,
 															SwingComponentModels componentModels) {
 		super(new SwingEntityEditor(entityType, connectionProvider, componentModels));
-		editor().events().persisted().addListener(() -> editor().comboBoxModels().column().values()
-						.forEach(comboBoxModel -> comboBoxModel.items().refresh()));
 	}
 }